expose container configuraotr
